Improves tools behind the --changed flag
