+++ test/helpers/automation_tools/utilities.py
@@ -2,7 +2,6 @@ from __future__ import print_function
 from contextlib import contextmanager
 import subprocess, os, sys
 
-
 def get_files(directory):
     print(directory)
     for _, _, files in os.walk(directory):
@@ -58,3 +57,29 @@ def get_output(commands):
         print(e)
         output = e.output
     return output.decode('utf-8')
+
+# return a list of the immediate subdirectories
+def get_package_names():
+    _, packageNames, _ = next(os.walk(os.getcwd()))
+    return packageNames
+
+# return a dict of filenames that correspond to full paths
+def get_file_names():
+    fileNames = {}
+    for subdir, _, files in os.walk(os.getcwd()):
+        for filename in files:
+            if filename.endswith('.py') and filename.startswith('test'):
+                fileNames[filename] = subdir + os.sep + filename
+    return fileNames
+
+def get_changed_files():
+    output = get_output(['git', 'diff', '--name-only'])
+    return output.splitlines()
+
+def get_changed_core_files(core='pygsti'):
+    return (filename.split(core, 1)[1] for filename in get_changed_files() if core in filename)
+
+def get_changed_packages():
+    return (corefile.split('/')[1] for corefile in get_changed_core_files())
+
+
